# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloudberry Health API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PatientSignupInputPrimaryGoal =
  (typeof PatientSignupInputPrimaryGoal)[keyof typeof PatientSignupInputPrimaryGoal];

export const PatientSignupInputPrimaryGoal = {
  weight_loss: "weight_loss",
  diabetes_management: "diabetes_management",
  both: "both",
} as const;

export interface PatientSignupInput {
  fullName: string;
  phone: string;
  city: string;
  primaryGoal: PatientSignupInputPrimaryGoal;
  /** @nullable */
  preferredCallbackTime?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  selectedPlan?: string | null;
}

export interface PatientSigninInput {
  phone: string;
  password: string;
}

export interface StaffSigninInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  role: string;
  userId?: number;
}

export interface LeadInput {
  fullName: string;
  phone: string;
  city: string;
  primaryGoal: string;
  /** @nullable */
  preferredCallbackTime?: string | null;
  /** @nullable */
  email?: string | null;
}

export interface Lead {
  id: number;
  fullName: string;
  phone: string;
  city: string;
  primaryGoal: string;
  /** @nullable */
  preferredCallbackTime?: string | null;
  /** @nullable */
  email?: string | null;
  createdAt: string;
}

export interface PhysicianLeadInput {
  name: string;
  specialty: string;
  /** @nullable */
  clinicOrHospital?: string | null;
  city: string;
  phone: string;
  /** @nullable */
  preferredCallbackTime?: string | null;
}

export interface PhysicianLead {
  id: number;
  name: string;
  specialty: string;
  /** @nullable */
  clinicOrHospital?: string | null;
  city: string;
  phone: string;
  /** @nullable */
  preferredCallbackTime?: string | null;
  createdAt: string;
}

export interface Patient {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  city: string;
  primaryGoal: string;
  plan: string;
  weekNumber: number;
  /** @nullable */
  startingWeight?: number | null;
  /** @nullable */
  currentWeight?: number | null;
  /** @nullable */
  targetWeight?: number | null;
  createdAt: string;
}

export interface PatientDashboard {
  patient: Patient;
  weekNumber: number;
  todayGoals: string[];
  /** @nullable */
  weightChange: number | null;
  /** @nullable */
  glucoseScore: number | null;
  /** @nullable */
  nutritionAdherence: number | null;
  /** @nullable */
  activityAdherence: number | null;
  /** @nullable */
  mood?: string | null;
  streak: number;
  /** @nullable */
  nextConsultation?: string | null;
  /** @nullable */
  medicationNote?: string | null;
}

export type CheckinInputMealsFollowed =
  (typeof CheckinInputMealsFollowed)[keyof typeof CheckinInputMealsFollowed];

export const CheckinInputMealsFollowed = {
  yes: "yes",
  partially: "partially",
  no: "no",
} as const;

export type CheckinInputEnergyLevel =
  (typeof CheckinInputEnergyLevel)[keyof typeof CheckinInputEnergyLevel];

export const CheckinInputEnergyLevel = {
  low: "low",
  moderate: "moderate",
  good: "good",
} as const;

export type CheckinInputMood =
  (typeof CheckinInputMood)[keyof typeof CheckinInputMood];

export const CheckinInputMood = {
  stressed: "stressed",
  neutral: "neutral",
  positive: "positive",
} as const;

export interface CheckinInput {
  mealsFollowed: CheckinInputMealsFollowed;
  activityCompleted: boolean;
  energyLevel: CheckinInputEnergyLevel;
  mood: CheckinInputMood;
  /** @nullable */
  glucoseReading?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface Checkin {
  id: number;
  mealsFollowed: string;
  activityCompleted: boolean;
  energyLevel: string;
  mood: string;
  /** @nullable */
  glucoseReading?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type MetricInputType =
  (typeof MetricInputType)[keyof typeof MetricInputType];

export const MetricInputType = {
  weight: "weight",
  glucose_fasting: "glucose_fasting",
  glucose_postmeal: "glucose_postmeal",
  steps: "steps",
} as const;

export interface MetricInput {
  type: MetricInputType;
  value: number;
  date: string;
  /** @nullable */
  notes?: string | null;
}

export interface MetricEntry {
  id: number;
  type: string;
  value: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface Tip {
  id: number;
  title: string;
  body: string;
  category: string;
  /** @nullable */
  icon?: string | null;
}

export interface AppointmentInput {
  careTeamMember: string;
  preferredDate: string;
  /** @nullable */
  notes?: string | null;
}

export interface RescheduleInput {
  newDate: string;
}

export interface Appointment {
  id: number;
  careTeamMember: string;
  role: string;
  scheduledAt: string;
  status: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export type CoachPatientRowStatus =
  (typeof CoachPatientRowStatus)[keyof typeof CoachPatientRowStatus];

export const CoachPatientRowStatus = {
  active: "active",
  paused: "paused",
  completed: "completed",
} as const;

export type CoachPatientRowRiskLevel =
  (typeof CoachPatientRowRiskLevel)[keyof typeof CoachPatientRowRiskLevel];

export const CoachPatientRowRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CoachPatientRow {
  id: number;
  fullName: string;
  plan?: string;
  status: CoachPatientRowStatus;
  riskLevel: CoachPatientRowRiskLevel;
  /** @nullable */
  lastCheckinAt: string | null;
  /** @nullable */
  nextSessionAt?: string | null;
  adherencePct: number;
  weekNumber?: number;
}

export interface PatientNote {
  id: number;
  content: string;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface PatientPlan {
  id: number;
  patientId: number;
  /** @nullable */
  nutritionPlan?: string | null;
  /** @nullable */
  activityPlan?: string | null;
  /** @nullable */
  weeklyGoals?: string | null;
  updatedAt?: string;
}

export interface CoachPatientDetail {
  patient: Patient;
  checkins: Checkin[];
  metrics: MetricEntry[];
  notes: PatientNote[];
  plan: PatientPlan;
}

export interface NoteInput {
  content: string;
  /** @nullable */
  category?: string | null;
}

export interface PlanUpdate {
  /** @nullable */
  nutritionPlan?: string | null;
  /** @nullable */
  activityPlan?: string | null;
  /** @nullable */
  weeklyGoals?: string | null;
}

export interface OpsDashboard {
  activePatients: number;
  dailyAdherencePct: number;
  missedCheckins: number;
  highRiskCount: number;
  upcomingAppointments: number;
  escalationsPending: number;
  totalLeads?: number;
  conversionRate?: number;
}

export interface OpsPatientRow {
  id: number;
  fullName: string;
  plan: string;
  status: string;
  riskLevel: string;
  adherencePct: number;
  weekNumber: number;
  /** @nullable */
  lastCheckinAt?: string | null;
  /** @nullable */
  assignedCoach?: string | null;
  escalated?: boolean;
}

export interface AssignCoachInput {
  coachId: number;
}

export interface EscalationInput {
  reason: string;
}