# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cloudberry Health API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/patient/signup:
    post:
      operationId: patientSignup
      tags: [auth]
      summary: Patient sign up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientSignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/patient/signin:
    post:
      operationId: patientSignin
      tags: [auth]
      summary: Patient sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientSigninInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/coach/signin:
    post:
      operationId: coachSignin
      tags: [auth]
      summary: Coach/Nutritionist sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffSigninInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/ops/signin:
    post:
      operationId: opsSignin
      tags: [auth]
      summary: Operations team sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffSigninInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /leads:
    post:
      operationId: submitLead
      tags: [leads]
      summary: Submit a patient lead from the public marketing page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/physician:
    post:
      operationId: submitPhysicianLead
      tags: [leads]
      summary: Submit a physician partnership enquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhysicianLeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicianLead"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: Get upcoming appointments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: bookAppointment
      tags: [appointments]
      summary: Book or request an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}/reschedule:
    patch:
      operationId: rescheduleAppointment
      tags: [appointments]
      summary: Reschedule an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RescheduleInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
  /coach/patients/{id}:
    get:
      operationId: getCoachPatientDetail
      tags: [coaches]
      summary: Get full patient details for coach
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachPatientDetail"
  /coach/patients/{id}/notes:
    post:
      operationId: addPatientNote
      tags: [coaches]
      summary: Add a note to patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientNote"
  /coach/patients/{id}/plan:
    patch:
      operationId: updatePatientPlan
      tags: [coaches]
      summary: Update patient nutrition/activity plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientPlan"
...
components:
  schemas:
...
    PatientSignupInput:
      type: object
      required: [fullName, phone, city, primaryGoal]
      properties:
        fullName:
          type: string
        phone:
          type: string
        city:
          type: string
        primaryGoal:
          type: string
          enum: [weight_loss, diabetes_management, both]
        preferredCallbackTime:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        selectedPlan:
          type: ["string", "null"]
    PatientSigninInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    StaffSigninInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, role]
      properties:
        token:
          type: string
        role:
          type: string
        userId:
          type: integer
    LeadInput:
      type: object
      required: [fullName, phone, city, primaryGoal]
      properties:
        fullName:
          type: string
        phone:
          type: string
        city:
          type: string
        primaryGoal:
          type: string
        preferredCallbackTime:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
    Lead:
      type: object
      required: [id, fullName, phone, city, primaryGoal, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        city:
          type: string
        primaryGoal:
          type: string
        preferredCallbackTime:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
    PhysicianLeadInput:
      type: object
      required: [name, specialty, city, phone]
      properties:
        name:
          type: string
        specialty:
          type: string
        clinicOrHospital:
          type: ["string", "null"]
        city:
          type: string
        phone:
          type: string
        preferredCallbackTime:
          type: ["string", "null"]
    PhysicianLead:
      type: object
      required: [id, name, specialty, city, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        clinicOrHospital:
          type: ["string", "null"]
        city:
          type: string
        phone:
          type: string
        preferredCallbackTime:
          type: ["string", "null"]
        createdAt:
          type: string
    Patient:
      type: object
      required: [id, fullName, phone, city, primaryGoal, plan, weekNumber, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        city:
          type: string
        primaryGoal:
          type: string
        plan:
          type: string
        weekNumber:
          type: integer
        startingWeight:
          type: ["number", "null"]
        currentWeight:
          type: ["number", "null"]
        targetWeight:
          type: ["number", "null"]
        createdAt:
          type: string
...
    Checkin:
      type: object
      required: [id, mealsFollowed, activityCompleted, energyLevel, mood, createdAt]
      properties:
        id:
          type: integer
        mealsFollowed:
          type: string
        activityCompleted:
          type: boolean
        energyLevel:
          type: string
        mood:
          type: string
        glucoseReading:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    MetricEntry:
      type: object
      required: [id, type, value, date, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        value:
          type: number
        date:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    AppointmentInput:
      type: object
      required: [careTeamMember, preferredDate]
      properties:
        careTeamMember:
          type: string
        preferredDate:
          type: string
        notes:
          type: ["string", "null"]
    RescheduleInput:
      type: object
      required: [newDate]
      properties:
        newDate:
          type: string
    Appointment:
      type: object
      required: [id, careTeamMember, role, scheduledAt, status]
      properties:
        id:
          type: integer
        careTeamMember:
          type: string
        role:
          type: string
        scheduledAt:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
...
    CoachPatientDetail:
      type: object
      required: [patient, checkins, metrics, notes, plan]
      properties:
        patient:
          $ref: "#/components/schemas/Patient"
        checkins:
          type: array
          items:
            $ref: "#/components/schemas/Checkin"
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricEntry"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/PatientNote"
        plan:
          $ref: "#/components/schemas/PatientPlan"
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        category:
          type: ["string", "null"]
    PatientNote:
      type: object
      required: [id, content, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        category:
          type: ["string", "null"]
        createdAt:
          type: string
    PlanUpdate:
      type: object
      properties:
        nutritionPlan:
          type: ["string", "null"]
        activityPlan:
          type: ["string", "null"]
        weeklyGoals:
          type: ["string", "null"]
    PatientPlan:
      type: object
      required: [id, patientId]
      properties:
        id:
          type: integer
        patientId:
          type: integer
        nutritionPlan:
          type: ["string", "null"]
        activityPlan:
          type: ["string", "null"]
        weeklyGoals:
          type: ["string", "null"]
        updatedAt:
          type: string
...