# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloudberry Health API
 * OpenAPI spec version: 0.1.0
 */

export interface Checkin {
  id: number;
  mealsFollowed: string;
  activityCompleted: boolean;
  energyLevel: string;
  mood: string;
  /** @nullable */
  glucoseReading?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}