# lib/api-zod/src/generated/types/metricEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloudberry Health API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricEntry {
  id: number;
  type: string;
  value: number;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}