# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloudberry Health API
 * OpenAPI spec version: 0.1.0
 */

export interface Patient {
  id: number;
  fullName: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  city: string;
  primaryGoal: string;
  plan: string;
  weekNumber: number;
  /** @nullable */
  startingWeight?: number | null;
  /** @nullable */
  currentWeight?: number | null;
  /** @nullable */
  targetWeight?: number | null;
  createdAt: string;
}