# lib/api-zod/src/generated/types/physicianLeadInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cloudberry Health API
 * OpenAPI spec version: 0.1.0
 */

export interface PhysicianLeadInput {
  name: string;
  specialty: string;
  /** @nullable */
  clinicOrHospital?: string | null;
  city: string;
  phone: string;
  /** @nullable */
  preferredCallbackTime?: string | null;
}